+++ packages/amplify-ui-components/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.9.34](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/ui-components@1.9.33...@aws-amplify/ui-components@1.9.34) (2022-09-20)
+
+**Note:** Version bump only for package @aws-amplify/ui-components
+
+
+
+
+
 ## [1.9.33](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/ui-components@1.9.32...@aws-amplify/ui-components@1.9.33) (2022-09-08)
 
 **Note:** Version bump only for package @aws-amplify/ui-components
