+++ packages/amplify-ui-react/package.json
@@ -2,7 +2,7 @@
 	"name": "@aws-amplify/ui-react",
 	"private": "true",
 	"sideEffects": false,
-	"version": "1.2.53",
+	"version": "1.2.54",
 	"description": "React specific wrapper for @aws-amplify/ui-components",
 	"publishConfig": {
 		"access": "public"
@@ -33,7 +33,7 @@
 		"typescript": "^3.3.4000"
 	},
 	"dependencies": {
-		"@aws-amplify/ui-components": "1.9.33"
+		"@aws-amplify/ui-components": "1.9.34"
 	},
 	"peerDependencies": {
 		"react": ">= 16.7.0",
