+++ packages/amplify-ui-storybook/package.json
@@ -1,6 +1,6 @@
 {
 	"name": "@aws-amplify/ui-storybook",
-	"version": "2.0.53",
+	"version": "2.0.54",
 	"private": true,
 	"dependencies": {
 		"@aws-amplify/ui-react": "0.2.38",
@@ -10,7 +10,7 @@
 		"@types/node": "^12.0.0",
 		"@types/react": "^16.9.0",
 		"@types/react-dom": "^16.9.0",
-		"aws-amplify": "4.3.35",
+		"aws-amplify": "4.3.36",
 		"react": "^16.12.0",
 		"react-app-polyfill": "^1.0.6",
 		"react-dom": "^16.12.0",
