+++ packages/amplify-ui-vue/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.48](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/ui-vue@1.1.47...@aws-amplify/ui-vue@1.1.48) (2022-09-20)
+
+**Note:** Version bump only for package @aws-amplify/ui-vue
+
+
+
+
+
 ## [1.1.47](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/ui-vue@1.1.46...@aws-amplify/ui-vue@1.1.47) (2022-09-08)
 
 **Note:** Version bump only for package @aws-amplify/ui-vue
