+++ packages/analytics/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.2.21](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/analytics@5.2.20...@aws-amplify/analytics@5.2.21) (2022-09-20)
+
+**Note:** Version bump only for package @aws-amplify/analytics
+
+
+
+
+
 ## [5.2.20](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/analytics@5.2.19...@aws-amplify/analytics@5.2.20) (2022-09-08)
 
 **Note:** Version bump only for package @aws-amplify/analytics
