+++ packages/api-graphql/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.18](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api-graphql@2.3.17...@aws-amplify/api-graphql@2.3.18) (2022-09-20)
+
+**Note:** Version bump only for package @aws-amplify/api-graphql
+
+
+
+
+
 ## [2.3.17](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api-graphql@2.3.16...@aws-amplify/api-graphql@2.3.17) (2022-09-08)
 
 **Note:** Version bump only for package @aws-amplify/api-graphql
