+++ packages/api-rest/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@aws-amplify/api-rest",
-  "version": "2.0.53",
+  "version": "2.0.54",
   "description": "Api-rest category of aws-amplify",
   "main": "./lib/index.js",
   "module": "./lib-esm/index.js",
@@ -41,7 +41,7 @@
   },
   "homepage": "https://aws-amplify.github.io/",
   "dependencies": {
-    "@aws-amplify/core": "4.7.4",
+    "@aws-amplify/core": "4.7.5",
     "axios": "0.26.0"
   },
   "jest": {
