+++ packages/auth/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@aws-amplify/auth",
-  "version": "4.6.6",
+  "version": "4.6.7",
   "description": "Auth category of aws-amplify",
   "main": "./lib/index.js",
   "module": "./lib-esm/index.js",
@@ -41,8 +41,8 @@
   },
   "homepage": "https://aws-amplify.github.io/",
   "dependencies": {
-    "@aws-amplify/cache": "4.0.55",
-    "@aws-amplify/core": "4.7.4",
+    "@aws-amplify/cache": "4.0.56",
+    "@aws-amplify/core": "4.7.5",
     "amazon-cognito-identity-js": "5.2.10",
     "crypto-js": "^4.1.1"
   },
