+++ packages/aws-amplify-angular/package.json
@@ -1,7 +1,7 @@
 {
 	"name": "aws-amplify-angular",
 	"private": "true",
-	"version": "6.0.53",
+	"version": "6.0.54",
 	"description": "AWS Amplify Angular Components",
 	"main": "bundles/aws-amplify-angular.umd.js",
 	"module": "dist/index.js",
@@ -39,7 +39,7 @@
 		"@types/zen-observable": "^0.5.3",
 		"angular2-template-loader": "^0.6.2",
 		"awesome-typescript-loader": "^4.0.1",
-		"aws-amplify": "4.3.35",
+		"aws-amplify": "4.3.36",
 		"babel-core": "^6.26.3",
 		"babel-plugin-lodash": "^3.3.4",
 		"babel-preset-env": "^1.7.0",
