+++ packages/aws-amplify/package.json
@@ -1,6 +1,6 @@
 {
   "name": "aws-amplify",
-  "version": "4.3.35",
+  "version": "4.3.36",
   "description": "AWS Amplify is a JavaScript library for Frontend and mobile developers building cloud-enabled applications.",
   "main": "./lib/index.js",
   "module": "./lib-esm/index.js",
@@ -34,19 +34,19 @@
   },
   "homepage": "https://aws-amplify.github.io/",
   "dependencies": {
-    "@aws-amplify/analytics": "5.2.20",
-    "@aws-amplify/api": "4.0.53",
-    "@aws-amplify/auth": "4.6.6",
-    "@aws-amplify/cache": "4.0.55",
-    "@aws-amplify/core": "4.7.4",
-    "@aws-amplify/datastore": "3.12.10",
-    "@aws-amplify/geo": "1.3.16",
-    "@aws-amplify/interactions": "4.1.1",
-    "@aws-amplify/predictions": "4.0.53",
-    "@aws-amplify/pubsub": "4.5.3",
-    "@aws-amplify/storage": "4.5.6",
+    "@aws-amplify/analytics": "5.2.21",
+    "@aws-amplify/api": "4.0.54",
+    "@aws-amplify/auth": "4.6.7",
+    "@aws-amplify/cache": "4.0.56",
+    "@aws-amplify/core": "4.7.5",
+    "@aws-amplify/datastore": "3.12.11",
+    "@aws-amplify/geo": "1.3.17",
+    "@aws-amplify/interactions": "4.1.2",
+    "@aws-amplify/predictions": "4.0.54",
+    "@aws-amplify/pubsub": "4.5.4",
+    "@aws-amplify/storage": "4.5.7",
     "@aws-amplify/ui": "2.0.5",
-    "@aws-amplify/xr": "3.0.53"
+    "@aws-amplify/xr": "3.0.54"
   },
   "jest": {
     "globals": {
