+++ packages/core/package.json
@@ -1,6 +1,6 @@
 {
 	"name": "@aws-amplify/core",
-	"version": "4.7.4",
+	"version": "4.7.5",
 	"description": "Core category of aws-amplify",
 	"main": "./lib/index.js",
 	"module": "./lib-esm/index.js",
