+++ packages/core/src/Platform/version.ts
@@ -1,2 +1,2 @@
 // generated by genversion
-export const version = '4.7.4';
+export const version = '4.7.5';
