+++ packages/datastore-storage-adapter/package.json
@@ -1,6 +1,6 @@
 {
 	"name": "@aws-amplify/datastore-storage-adapter",
-	"version": "1.3.13",
+	"version": "1.3.14",
 	"description": "SQLite storage adapter for Amplify DataStore ",
 	"main": "./lib/index.js",
 	"module": "./lib-esm/index.js",
@@ -34,8 +34,8 @@
 	},
 	"homepage": "https://aws-amplify.github.io/",
 	"devDependencies": {
-		"@aws-amplify/core": "4.7.4",
-		"@aws-amplify/datastore": "3.12.10",
+		"@aws-amplify/core": "4.7.5",
+		"@aws-amplify/datastore": "3.12.11",
 		"@types/react-native-sqlite-storage": "5.0.1",
 		"expo-file-system": "13.1.4",
 		"expo-sqlite": "10.1.0",
