+++ packages/datastore/package.json
@@ -1,6 +1,6 @@
 {
 	"name": "@aws-amplify/datastore",
-	"version": "3.12.10",
+	"version": "3.12.11",
 	"description": "AppSyncLocal support for aws-amplify",
 	"main": "./lib/index.js",
 	"module": "./lib-esm/index.js",
@@ -48,10 +48,10 @@
 		"fake-indexeddb": "3.0.0"
 	},
 	"dependencies": {
-		"@aws-amplify/api": "4.0.53",
-		"@aws-amplify/auth": "4.6.6",
-		"@aws-amplify/core": "4.7.4",
-		"@aws-amplify/pubsub": "4.5.3",
+		"@aws-amplify/api": "4.0.54",
+		"@aws-amplify/auth": "4.6.7",
+		"@aws-amplify/core": "4.7.5",
+		"@aws-amplify/pubsub": "4.5.4",
 		"amazon-cognito-identity-js": "5.2.10",
 		"idb": "5.0.6",
 		"immer": "9.0.6",
