+++ packages/geo/package.json
@@ -1,6 +1,6 @@
 {
 	"name": "@aws-amplify/geo",
-	"version": "1.3.16",
+	"version": "1.3.17",
 	"description": "Geo category for aws-amplify",
 	"main": "./lib/index.js",
 	"module": "./lib-esm/index.js",
@@ -41,7 +41,7 @@
 	},
 	"homepage": "https://aws-amplify.github.io/",
 	"dependencies": {
-		"@aws-amplify/core": "4.7.4",
+		"@aws-amplify/core": "4.7.5",
 		"@aws-sdk/client-location": "3.48.0",
 		"@turf/boolean-clockwise": "6.5.0",
 		"camelcase-keys": "6.2.2"
