+++ packages/pushnotification/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.3.33](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/pushnotification@4.3.32...@aws-amplify/pushnotification@4.3.33) (2022-09-20)
+
+**Note:** Version bump only for package @aws-amplify/pushnotification
+
+
+
+
+
 ## [4.3.32](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/pushnotification@4.3.31...@aws-amplify/pushnotification@4.3.32) (2022-09-08)
 
 **Note:** Version bump only for package @aws-amplify/pushnotification
