+++ packages/storage/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.5.7](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/storage@4.5.6...@aws-amplify/storage@4.5.7) (2022-09-20)
+
+**Note:** Version bump only for package @aws-amplify/storage
+
+
+
+
+
 ## [4.5.6](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/storage@4.5.5...@aws-amplify/storage@4.5.6) (2022-09-08)
 
 **Note:** Version bump only for package @aws-amplify/storage
