+++ packages/storage/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@aws-amplify/storage",
-  "version": "4.5.6",
+  "version": "4.5.7",
   "description": "Storage category of aws-amplify",
   "main": "./lib/index.js",
   "module": "./lib-esm/index.js",
@@ -41,7 +41,7 @@
   },
   "homepage": "https://aws-amplify.github.io/",
   "dependencies": {
-    "@aws-amplify/core": "4.7.4",
+    "@aws-amplify/core": "4.7.5",
     "@aws-sdk/client-s3": "3.6.1",
     "@aws-sdk/s3-request-presigner": "3.6.1",
     "@aws-sdk/util-create-request": "3.6.1",
